pyproject.toml
[tool.poetry]
name = "svd"
version = "0.1.0"
description = ""
authors = ["Evgeniy Trubnikov <parmesanshees@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
pillow = "^11.0.0"
pathlib = "^1.0.1"
rich = "^13.9.4"
matplotlib = "^3.9.2"

[tool.poetry.scripts]
main = 'svd.__init__:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
